Factor out FlashImage.
Fixes <https://github.com/oxidecomputer/amd-host-image-builder/issues/110>.
