+++ Makefile
@@ -3,7 +3,9 @@ all: milan-ethanol-x rome-ethanol-x milan-gimlet-b
 .PHONY: all clean tests
 .PHONY: FRC
 
-SOURCES = amd-host-image-builder-config/src/lib.rs src/hole.rs src/main.rs src/static_config.rs Cargo.toml amd-host-image-builder-config/Cargo.toml Cargo.lock
+SOURCES = amd-host-image-builder-config/src/lib.rs \
+    src/hole.rs src/main.rs src/static_config.rs src/images.rs src/allocators.rs \
+    Cargo.toml amd-host-image-builder-config/Cargo.toml Cargo.lock
 
 CARGO = cargo
