+++ src/main.rs
@@ -8,18 +8,13 @@ use amd_host_image_builder_config::{
     SerdePspDirectoryVariant, SerdePspEntrySource,
     TryFromSerdeDirectoryEntryWithContext,
 };
-use core::cell::RefCell;
 use core::convert::TryFrom;
 use core::convert::TryInto;
 use static_assertions::const_assert;
 use std::fs;
 use std::fs::File;
-use std::fs::OpenOptions;
 use std::io::BufReader;
 use std::io::Read;
-use std::io::Seek;
-use std::io::SeekFrom;
-use std::io::Write;
 use std::path::Path;
 use std::path::PathBuf;
 use structopt::StructOpt;
@@ -49,120 +44,8 @@ fn test_bitfield_serde() {
 mod hole;
 use hole::Hole;
 
-struct FlashImage {
-    file: RefCell<File>,
-    filename: PathBuf,
-    erasable_block_size: usize,
-    buffer: RefCell<Vec<u8>>,
-}
-
-impl FlashRead for FlashImage {
-    fn read_exact(
-        &self,
-        location: Location,
-        buffer: &mut [u8],
-    ) -> amd_flash::Result<()> {
-        let mut file = self.file.borrow_mut();
-        file.seek(SeekFrom::Start(location.into())).map_err(|e| {
-            let filename = &self.filename;
-            eprintln!("Error seeking in flash image {filename:?}: {e:?}",);
-            amd_flash::Error::Io
-        })?;
-        file.read_exact(buffer).map_err(|e| {
-            let filename = &self.filename;
-            eprintln!("Error reading from flash image {filename:?}: {e:?}",);
-            amd_flash::Error::Io
-        })
-    }
-}
-
-impl FlashAlign for FlashImage {
-    fn erasable_block_size(&self) -> usize {
-        self.erasable_block_size
-    }
-}
-
-impl FlashWrite for FlashImage {
-    fn erase_block(&self, location: ErasableLocation) -> amd_flash::Result<()> {
-        let erasable_block_size = self.erasable_block_size();
-        let location = self.location(location)?;
-        let mut file = self.file.borrow_mut();
-        match file.seek(SeekFrom::Start(location.into())) {
-            Ok(_) => {}
-            Err(e) => {
-                let filename = &self.filename;
-                eprintln!("Error seeking in flash image {filename:?}: {e:?}",);
-                return Err(amd_flash::Error::Io);
-            }
-        }
-        let mut buffer = self.buffer.borrow_mut();
-        buffer.fill(0xff);
-        match file.write(&buffer[..]) {
-            Ok(size) => {
-                assert!(size == erasable_block_size);
-                Ok(())
-            }
-            Err(e) => {
-                eprintln!(
-                    "Error writing to flash image {:?}: {:?}",
-                    self.filename, e
-                );
-                Err(amd_flash::Error::Io)
-            }
-        }
-    }
-    fn erase_and_write_block(
-        &self,
-        location: ErasableLocation,
-        buffer: &[u8],
-    ) -> amd_flash::Result<()> {
-        let erasable_block_size = self.erasable_block_size;
-        if buffer.len() > erasable_block_size {
-            return Err(amd_flash::Error::Programmer);
-        }
-        let location = self.location(location)?;
-        let mut file = self.file.borrow_mut();
-        file.seek(SeekFrom::Start(location.into())).map_err(|e| {
-            eprintln!(
-                "Error seeking in flash image {:?}: {:?}",
-                self.filename, e
-            );
-            amd_flash::Error::Io
-        })?;
-        let mut xbuffer = self.buffer.borrow_mut();
-        xbuffer.fill(0xff);
-        xbuffer[..buffer.len()].copy_from_slice(buffer);
-        match file.write(&xbuffer) {
-            Ok(size) => {
-                assert!(size == xbuffer.len());
-                Ok(())
-            }
-            Err(e) => {
-                eprintln!(
-                    "Error writing to flash image {:?}: {:?}",
-                    self.filename, e
-                );
-                Err(amd_flash::Error::Io)
-            }
-        }
-    }
-}
-
-impl FlashImage {
-    fn new(file: File, filename: &Path, erasable_block_size: usize) -> Self {
-        assert!(erasable_block_size.is_power_of_two());
-        Self {
-            file: RefCell::new(file),
-            filename: filename.to_path_buf(),
-            erasable_block_size,
-            buffer: RefCell::new(
-                std::iter::repeat(0xff).take(erasable_block_size).collect(),
-            ),
-        }
-    }
-}
-
-type AlignedLocation = ErasableLocation;
+mod images;
+use images::FlashImage;
 
 /// Open SOURCE_FILENAME and checks its size.
 /// If TARGET_SIZE is given, make sure the file is at most as big as that.
@@ -625,27 +508,9 @@ fn run() -> std::io::Result<()> {
             )
         };
 
-    let file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .create(true)
-        .open(filename)?;
-    file.set_len(static_config::IMAGE_SIZE.into())?;
     const ERASABLE_BLOCK_SIZE: usize = 0x1000;
     const_assert!(ERASABLE_BLOCK_SIZE.is_power_of_two());
-    let storage = FlashImage::new(file, filename, ERASABLE_BLOCK_SIZE);
-    let erasable_block_size = storage.erasable_block_size();
-    let mut position: AlignedLocation = storage
-        .erasable_location(0)
-        .ok_or(amd_flash::Error::Alignment)
-        .map_err(flash_to_io_error)?;
-    while Location::from(position) < static_config::IMAGE_SIZE {
-        FlashWrite::erase_block(&storage, position)
-            .map_err(flash_to_io_error)?;
-        position =
-            position.advance(erasable_block_size).map_err(flash_to_io_error)?;
-    }
-    assert!(Location::from(position) == static_config::IMAGE_SIZE);
+    let storage = FlashImage::create(filename, ERASABLE_BLOCK_SIZE)?;
     let path = Path::new(&opts.efs_configuration_filename);
     let data = std::fs::read_to_string(path)?;
     let config: SerdeConfig =
